+++ resources/helpers/TranslatorHelper.php
@@ -1,0 +1,7 @@
+<?php
+namespace App\Helpers;
+
+interface TranslatorHelper {
+    public function trans(string $key): string;
+    public function trans_choice(string $key, int $number): string;
+}