+++ templates/database/seeders/PageSeeder.php
@@ -6,9 +6,10 @@ use App\Models\Module;
 use App\Models\Menuitem;
 use App\Models\SProperty;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 
-class PageSeeder extends SeederS
+class PageSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -24,6 +25,17 @@ class PageSeeder extends SeederS
 
         SProperty::insertIfNotExists(1141, 'pagetype', 'menu', 10, 'MN');
         SProperty::insertIfNotExists(1142, 'pagetype', 'help', 20, 'HP');
+
+        DB::table('pages')->insert([
+            'title' => 'Startmenu',
+            'name' => 'main',
+            'contents' => "== Menu ==\n---- %col%\n== Hilfe ==\n: [[/page-showhelp/wikisyntax|Wiki-Syntax]]<br>\n:: [[/page-showhelp/wiki-absaetze|Wiki-Absätze]]<br>\n:: [[/page-showpretty/15|Wiki-Tabelle]]",
+            'pagetype_scope' => 1141,
+            'markup_scope' => 1122,
+            'order' => 1,
+            'language_scope' => 1202,
+            'owner_id' => 1,
+        ]);
     }
 }